+++ pyunicore/helpers/__init__.py
@@ -1,13 +1,3 @@
-from .requests import Credentials
-from .requests import Export
-from .requests import Import
-from .requests import JobDescription
-from .requests import Resources
-from .connect import create_transport
-from .connect import connect_to_site_from_registry
-from .connect import connect_to_site
-from .connect import connect_to_registry
-from .connect import UserAuthorization
-from .connect import TokenAuthorization
-from .connect import AuthorizationFailedException
-from .job_status import JobStatus
+from . import jobs
+from . import workflows
+from . import connect
