+++ pyunicore/helpers/jobs/__init__.py
@@ -1,6 +1,6 @@
 from .data import Credentials
 from .data import Export
 from .data import Import
-from .job_description import JobDescription
+from .description import Description
 from .resources import Resources
-from . import workflows
+from .status import Status
