+++ pyunicore/helpers/workflows/__init__.py
@@ -3,7 +3,7 @@
 See https://unicore-dev.zam.kfa-juelich.de/documentation/workflow-8.0.0/workflow-manual.html  # noqa
 
 """
-from . import activity as Activity
-from .workflow_description import WorkflowDescription
+from . import activities
+from .description import Description
 from .variable import Variable
 from .transition import Transition
