+++ pyunicore/helpers/workflows/activities/__init__.py
@@ -6,4 +6,4 @@ from .activity import Synchronize
 from .activity import Hold
 from .job import Job
 from .modify_variable import ModifyVariable
-from . import loops as Loops
+from . import loops
