+++ pyunicore/helpers/workflows/activities/loops/body.py
@@ -2,9 +2,9 @@ import dataclasses
 from typing import Dict
 from typing import List
 
-from pyunicore.helpers.requests import _api_object
-from pyunicore.helpers.requests.workflows import transition
-from pyunicore.helpers.requests.workflows.activity import activity
+from pyunicore.helpers import _api_object
+from pyunicore.helpers.workflows import transition
+from pyunicore.helpers.workflows.activities import activity
 
 
 @dataclasses.dataclass
