+++ pyunicore/helpers/workflows/activities/loops/for_loop.py
@@ -3,8 +3,8 @@ from typing import Dict
 from typing import List
 from typing import Optional
 
-from pyunicore.helpers.requests import _api_object
-from pyunicore.helpers.requests.workflows import variable
+from pyunicore.helpers import _api_object
+from pyunicore.helpers.workflows import variable
 from . import _loop
