+++ pyunicore/helpers/workflows/transition.py
@@ -2,7 +2,7 @@ import dataclasses
 from typing import Optional
 from typing import Dict
 
-from pyunicore.helpers.requests import _api_object
+from pyunicore.helpers import _api_object
 
 
 @dataclasses.dataclass
