+++ tests/helpers/workflows/test_workflow_description.py
@@ -1,9 +1,9 @@
-import pyunicore.helpers.requests.workflows as workflows
+import pyunicore.helpers.workflows as workflows
 
 
 class TestWorkflowDescription:
     def test_to_dict(self):
-        activities = [workflows.Activity.Start]
+        activities = [workflows.activities.Start(id="test-start")]
         transitions = [
             workflows.Transition(
                 from_="here",
@@ -18,13 +18,20 @@ class TestWorkflowDescription:
             )
         ]
 
-        workflow = workflows.WorkflowDescription(
+        workflow = workflows.Description(
             activities=activities,
             transitions=transitions,
             variables=variables,
         )
-        expected = {}
+        expected = {
+            "activities": [{"id": "test-start", "type": "START"}],
+            "transitions": [{"from": "here", "to": "there"}],
+            "variables": [
+                {"initial_value": 1, "name": "test-variable", "type": "INTEGER"}
+            ],
+        }
 
         result = workflow.to_dict()
 
+        breakpoint()
         assert result == expected
